Add Docker support
Add the ability to build a Docker image and run yams from a Docker
container.

To ensure that the final image is reasonably small, we leverage the
newer multistage build pattern introduced with Docker 17.05.0.  This
seems to leave around the intermediate results as untagged images -- an
issue we'll have to solve in the future.

Debian is used as the base of our run-image because that's what the rust
image uses.  We further use the slim variant to get a smaller image.

The final run-image uses tini as a lightweight init to handle signal
proxying correctly.  Without it, SIGINTs and SIGTERMs do not stop yams.